requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
Authlib==1.3.2
blinker==1.8.2
celery[redis]==5.3.1 
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dask==2.10.1
datashape==0.5.4
distributed==2.10.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
greenlet==3.0.3
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
jsonpointer==2.1
mccabe==0.6.1
mpmath==1.3.0
networkx==3.4.2
pillow==11.0.0
ply==3.11
protobuf==3.20.3
psycopg2==2.9.10
pycparser==2.22
pyls-spyder==0.4.0
pynvim==0.5.0
python-dotenv==1.0.1
redis==5.0.0  
requests==2.32.3
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchtext==0.18.0
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.3
